backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
transformers
torch
sentencepiece
langdetect
pydantic